시간복잡도 O(N^2) 판별 문제

+++ 시간복잡도/24264.cpp
@@ -8,13 +8,15 @@
 
 // 코드1 의 시간복잡도는 O(n^2) 이다. 즉, n번 반복함에 따라 코드1도 n^2번 실행되게 된다.
 
+// 단, n의 최댓값이 50만이므로 overflow 을 대비한다. long long으로 받는다.
+
 #include <iostream>
 
 using namespace std;
 
 int main(void) {
     ios_base::sync_with_stdio(false); cin.tie(nullptr);
-    int n;
+    long long n;
     cin >> n;
     cout << n*n << "\n";
     cout << "2";
